requirements.txt
llama-index==0.12.0
llama-index-vector-stores-chroma==0.4.1
llama-index-embeddings-openai==0.3.1
llama-index-llms-openai==0.3.6
llama-index-readers-file==0.4.2
chromadb==0.6.3
pypdf==4.3.1
python-docx==1.1.2
python-dotenv==1.0.1
rich==13.8.1
